// src/app/models/npro/IngredientDetails.ts
export interface IngredientDetail {
  ingredientId: string
  libraryId: string
  groupId: string
  foodCode: string
  foodName: string
  displayName: string
  description: string
  edibleProportion: number
  specificGravity: any
  totalSolids: any
  nitrogenConversionFactor: number
  glycerolConversionFactor: any
  costPerKilo: any
  internalReference: any
  liquid: any
}